more William's feedback
