+++ engine/src/cython/initialize.cpp
@@ -47,6 +47,7 @@
 #include "communication/CommunicationInterface/messageSender.hpp"
 #include "communication/CommunicationInterface/messageListener.hpp"
 #include "execution_graph/logic_controllers/taskflow/kernel.h"
+#include "execution_graph/logic_controllers/taskflow/executor.h"
 
 #include "error.hpp"
 
@@ -735,7 +736,6 @@ std::pair<std::pair<std::shared_ptr<CacheMachine>,std::shared_ptr<CacheMachine> 
 		ucp_context_h ucp_context = nullptr;
 		ucp_worker_h self_worker = nullptr;
 		if(protocol == comm::blazing_protocol::ucx){
-			std::cout<<"def wrong"<<std::endl;
 			ucp_context = reinterpret_cast<ucp_context_h>(workers_ucp_info[0].context_handle);
 
 			self_worker = CreatetUcpWorker(ucp_context);
