+++ engine/src/execution_graph/logic_controllers/taskflow/kernel.h
@@ -159,8 +159,8 @@ public:
 		return added;
 	}
 
-	/**.fetch_add(1, std::memory_order_relaxed);
-	 *
+	/**
+	 * @brief Adds a CacheData into the output cache.
 	 * @param cache_data The cache_data that will be added to the output cache.
 	 * @param cache_id The cache identifier.
 	 */
@@ -199,7 +199,7 @@ public:
 	 * @brief Adds a BlazingHostTable into the output cache.
 	 *
 	 * @param host_table The host table that will be added to the output cache.
-	 * @param cache_id The cache identifier..fetch_add(1, std::memory_order_relaxed);
+	 * @param cache_id The cache identifier.
 	 */
 	bool add_to_output_cache(std::unique_ptr<ral::frame::BlazingHostTable> host_table, std::string cache_id = "") {
 		CodeTimer cacheEventTimer(false);
@@ -325,108 +325,4 @@ public:
 
 
 }  // namespace cache
-
-
-namespace execution{
-
-class prioirity {
-public:
-
-private:
-	size_t priority_num_query; //can be used to prioritize one query over another
-	size_t priority_num_kernel; //can be 
-	
-};
-
-class executor;
-
-
-
-class task {
-public:
-
-
-	task(
-    std::vector<std::unique_ptr<ral::cache::CacheData > > inputs,
-    std::shared_ptr<ral::cache::CacheMachine> output,
-    size_t task_id,
-    ral::cache::kernel * kernel, size_t attempts_limit,
-    std::string kernel_process_name, size_t attempts = 0);
-
-	/**
-	* Function which runs the kernel process on the inputs and puts results into output.
-	* This function does not modify the inputs and can throw an exception. In the case it throws an exception it 
-	* gets placed back in the executor if it was a memory exception.
-	*/
-	void run(cudaStream_t stream, executor * executor);
-	void complete();
-protected:
-	std::vector<std::unique_ptr<ral::cache::CacheData > > inputs;	
-	std::shared_ptr<ral::cache::CacheMachine> output;
-	size_t task_id;
-	ral::cache::kernel * kernel;
-	size_t attempts = 0;
-	size_t attempts_limit;
-	std::string kernel_process_name = "";
-};
-
-
-
-class executor{
-public:
-	static executor * get_instance(){
-		if(_instance == nullptr){
-			throw std::runtime_error("Executor not initialized.");
-		}
-		return _instance;
-	}
-
-	static void init_executor(int num_threads){
-		if(!_instance){
-			_instance = new executor(num_threads);
-			auto thread = std::thread([_instance]{
-				_instance->execute();
-			});
-			thread.detach();
-		}
-
-	}
-
-	void execute();
-	size_t add_task(std::vector<std::unique_ptr<ral::cache::CacheData > > inputs,
-		std::shared_ptr<ral::cache::CacheMachine> output,
-		ral::cache::kernel * kernel,std::string kernel_process_name);
-
-	void add_task(std::vector<std::unique_ptr<ral::cache::CacheData > > inputs,
-		std::shared_ptr<ral::cache::CacheMachine> output,
-		ral::cache::kernel * kernel,
-		size_t attempts,
-		size_t task_id,std::string kernel_process_name);
-
-/*	
-	void add_task(ral::batch::DataSourceSequence * inputs,
-				std::shared_ptr<ral::cache::CacheMachine> output,
-		ral::cache::kernel * kernel,std::string kernel_process_name); 
-	
-	void add_task(ral::batch::DataSourceSe@rodquence * inputs,
-		std::shared_ptr<ral::cache::CacheMachine> output,
-		ral::cache::kernel * kernel,
-		size_t attempts,
-		size_t task_id,std::string kernel_process_name);
-*/
-
-private:
-	executor(int num_threads);
-	ctpl::thread_pool<BlazingThread> pool;
-	std::vector<cudaStream_t> streams; //one stream per thread
-	ral::cache::WaitingQueue< std::unique_ptr<task> > task_queue;
-	int shutdown = 0;
-	static executor * _instance;
-	std::atomic<int> task_id_counter;
-	size_t attempts_limit = 10;
-};
-
-
-} // namespace execution
-
 }  // namespace ral
